Jump tweaks
If player falls of a ledge, for a split second player logic is performed
as though the player were on the ground. This also fixed a bug where for
a frame the player wasn't on the ground when walking over slopes.
Pressing jump now causes the player to jump if it was pressed just
before hitting the ground